order_service/requirements.txt
Django==4.2
djangorestframework==3.14
requests==2.31
psycopg2==2.9.1

